added testing to views and models. added comments and doc strings

+++ application/app.py
@@ -8,7 +8,7 @@ import pygeohash as pgh
 from sqlalchemy.exc import IntegrityError, PendingRollbackError
 from validators import url as validate_url
 
-from models import db, connect_db, User, Artist, UserArtist, Event, UserEvent, WishList, CreateEvent
+from models import db, connect_db, User, Artist, UserArtist, Event, UserEvent, WishList
 from forms import NewUserForm, LoginForm, EditUserForm, ChangePasswordForm, ChangePfpForm
 from ticketmaster import TicketmasterAPI
 from spotify import SpotifyAPI
@@ -42,6 +42,8 @@ ticketmaster = TicketmasterAPI(api_key=TICKETMASTER_API_KEY)
 
 @app.before_request
 def add_user_to_g():
+    ''' adds currently logged in user to flask g for global use'''
+
     if CUR_U_ID in session:
         g.user = User.query.filter_by(id=session[CUR_U_ID]).first()
 
@@ -51,9 +53,14 @@ def add_user_to_g():
 
 @app.route('/')
 def homepage():
+    ''' returns homepage template based on if a user is logged in or if spotify is connected'''
+
+        # gets list of generic events
     generic_events = ticketmaster.get_generic_events()
+        # puts generic events in groups for carousel
     all_generic_events = [generic_events[0:5], generic_events[5:10], generic_events[10:15], generic_events[15:]]
 
+        # creates list of 5 generic artists for if spotify is not connected
     generic_artists = []
     for i in range(5):
         artist = Artist.query.filter_by(id=i+1).first()
@@ -61,7 +68,7 @@ def homepage():
 
     if not g.user:
         form = LoginForm()
-        return render_template('generic-homepage.html', form=form, all_events=all_generic_events)
+        return render_template('generic-homepage.html', form=form, all_events=all_generic_events, generic_artists=generic_artists)
     
     user = User.query.filter_by(username=g.user.username).first()
 
@@ -69,14 +76,16 @@ def homepage():
         zipcode = user.zipcode
         coords = get_lat_long(zipcode)
         geohash = get_geohash(coords)
+
+            # gets events based on users location
         generic_events_geohash = ticketmaster.get_generic_events(geohash=geohash)
+
+            # gets list of users wishlist
         wishlist = [event.event_id for event in user.wishlist]
 
         if session.get('spotify_token', None):
-
             artists = user.artists
             top_tracks = session['top_tracks']
-            print(top_tracks)
 
             if artists:
                 artist_num = 1
@@ -84,9 +93,13 @@ def homepage():
 
                 track_num = 1
                 top_tracks_numbered = []
+
+                    # creates numbered dict for top artists
                 for artist in artists[:5]:
                     top_artists.append({'id': artist_num, 'name': artist.name, 'img': artist.image})
                     artist_num += 1
+
+                    # creats numbered dict for top tracks
                 for track in top_tracks:
                     top_tracks_numbered.append({'id': track_num, 'name': track['name'], 'artist': track['artist'], 'img': track['image_url']})
                     track_num += 1
@@ -94,11 +107,14 @@ def homepage():
                 return render_template('user-homepage.html', user=user, all_events=all_generic_events, top_artists=top_artists, spot_login=True, generic_events_geohash=generic_events_geohash, wishlist=wishlist, top_tracks=top_tracks_numbered)
             
         return render_template('user-homepage.html', user=user, all_events=all_generic_events, generic_artists=generic_artists, generic_events_geohash=generic_events_geohash, wishlist=wishlist)
-    return render_template('generic-homepage.html', all_events=all_generic_events)
+    
+    return render_template('generic-homepage.html', all_events=all_generic_events, generic_artists=generic_artists)
 
 
 @app.route('/login', methods=['GET', 'POST'])
 def login():
+    ''' logs in a user using authentication. if authenticates, adds user to sessions current user'''
+
     form = LoginForm()
     if form.validate_on_submit():
         auth = User.authenticate(form.username.data, form.password.data)
@@ -114,12 +130,17 @@ def login():
 
 @app.route('/signup', methods=['GET', 'POST'])
 def signup():
+    ''' signs up a user, sets default profile image if not validated. '''
+
     form = NewUserForm()
 
     if form.validate_on_submit():
         try:
+                # gets data excluding csrf token
             data = {field: form[field].data for field in form._fields if field != 
             'csrf_token'}
+
+                # validates profile image from data
             if not validate_url(data['profile_img']):
                 data['profile_img'] = 'https://cdn-icons-png.freepik.com/512/5997/5997002.png'
 
@@ -137,32 +158,48 @@ def signup():
 
 @app.route('/user/details/<username>')
 def user_details(username):
+    ''' gets and displays user details'''
+
     if not g.user:
-        flash('You must be logged in to view this page.')
+        flash('You must be logged in to view this page.', 'danger')
         return redirect(url_for('homepage'))
     
     user = User.query.filter_by(username=username).first()
+    spot_login = True if session['spotify_token'] else False
 
-    return render_template('user-details.html', user=user)
+    return render_template('user-details.html', user=user, spot_login=spot_login)
 
 
 @app.route('/user/details/edit/<username>', methods=['GET', 'POST'])
 def edit_user(username):
-    u = User.query.filter_by(username=username).first()
+    ''' shows from to edit a user details'''
 
+    if not g.user:
+        flash('You must be logged in to view this page.', 'danger')
+        return redirect(url_for('homepage'))
+    
+    u = User.query.filter_by(username=username).first()
     form = EditUserForm(obj=u)
+
+    spot_login = True if session['spotify_token'] else False
 
     if form.validate_on_submit():
         try:
+                # gets data excluding csrf token
             data = {field: form[field].data for field in form._fields if field != 'csrf_token'}
             user_id = u.id
             update = User.update_details(user_id, **data)
             db.session.commit()
+
+            flash('Details Updated', 'success')
+            return redirect(f'/user/details/{u.username}')
             
         except PendingRollbackError:
             db.session.rollback()
             flash('An error occured', 'danger')
             return render_template('user-edit.html', form=form, u=u)
+        
+            # handles duplicate or taken email/usernames
         except IntegrityError as e:
             db.session.rollback()
             error_message = str(e.orig).lower()
@@ -172,14 +209,20 @@ def edit_user(username):
                 flash('Email already in use.', 'danger')
             return render_template('user-edit.html', form=form, u=u)
 
-    return render_template('user-edit.html', form=form, u=u)
+    return render_template('user-edit.html', form=form, user=u, spot_login=spot_login)
 
 
 @app.route('/user/password/edit/<username>', methods=['GET', 'POST'])
 def change_password(username):
-    u = User.query.filter_by(username=username).first()
+    ''' shows form to change a users password. '''
+
+    if not g.user:
+        flash('You must be logged in to view this page.', 'danger')
+        return redirect(url_for('homepage'))
     
+    u = User.query.filter_by(username=username).first()
     form = ChangePasswordForm()
+    spot_login = True if session['spotify_token'] else False
 
     if form.validate_on_submit():
         auth = User.authenticate(u.username, form.password.data)
@@ -195,17 +238,25 @@ def change_password(username):
         flash('Could not update password', 'danger')
         return redirect(f'/user/details/{u.username}')
     
-    return render_template('user-password-edit.html', form=form, u=u)
+    return render_template('user-password-edit.html', form=form, user=u, spot_login=spot_login)
 
 
 @app.route('/user/edit-pfp/<username>', methods=['GET', 'POST'])
 def change_pfp(username):
-    u = User.query.filter_by(username=username).first()
+    ''' shows form to update a users profile image '''
 
+    if not g.user:
+        flash('You must be logged in to view this page.', 'danger')
+        return redirect(url_for('homepage'))
+    
+    u = User.query.filter_by(username=username).first()
     form = ChangePfpForm()
+    spot_login = True if session['spotify_token'] else False
 
     if form.validate_on_submit():
         profile_img = form.profile_img.data
+
+            # validates profile image from data
         if not validate_url(profile_img):
             profile_img = 'https://cdn-icons-png.freepik.com/512/5997/5997002.png'
         update = User.update_pfp(u.username, profile_img)
@@ -218,24 +269,31 @@ def change_pfp(username):
         flash('could not update profile picture', 'danger')
         return redirect(f'/user/edit-pfp/{u.username}')
     
-    return render_template('user-pfp-edit.html', form=form, u=u)
+    return render_template('user-pfp-edit.html', form=form, user=u, spot_login=spot_login)
 
 
 @app.route('/add-to-wishlist/<event_id>', methods=['POST'])
 def add_to_wishlist(event_id):
+    ''' adds an event to wishlist by event id and currently logged in user'''
+
     if not g.user:
-        return None
+        flash('You must be logged in to view this page.', 'danger')
+        return redirect(url_for('homepage'))
+    
 
     user = User.query.filter_by(username=g.user.username).first()
 
+        # checks if event is in data base
     existing_event = Event.query.filter_by(event_id=event_id).first()
 
+        # adds event to data base if not in it
     if not existing_event:
         event = ticketmaster.get_event(event_id)
         event_to_add = Event(**event)
         db.session.add(event_to_add)
         db.session.commit()
 
+        # adds event to wishlist with users id
     new_wishlist = WishList(user_id=user.id, event_id=event_id)
     db.session.add(new_wishlist)
     db.session.commit()
@@ -245,6 +303,12 @@ def add_to_wishlist(event_id):
 
 @app.route('/remove-wishlist/<event_id>', methods=['POST'])
 def remove_from_wishlist(event_id):
+    ''' removes an event from wishlist '''
+
+    if not g.user:
+        flash('You must be logged in to view this page.', 'danger')
+        return redirect(url_for('homepage'))
+    
     user = User.query.filter_by(username=g.user.username).first()
     item = WishList.query.filter_by(user_id=user.id, event_id=event_id).first()
     db.session.delete(item)
@@ -255,6 +319,12 @@ def remove_from_wishlist(event_id):
 
 @app.route('/get-wishlist')
 def get_wishlist():
+    ''' returns a list of event ids based on users '''
+
+    if not g.user:
+        flash('You must be logged in to view this page.', 'danger')
+        return redirect(url_for('homepage'))
+    
     user = User.query.filter_by(username=g.user.username).first()
     wishlist_events = [event.event_id for event in user.wishlist]
 
@@ -263,12 +333,21 @@ def get_wishlist():
 
 @app.route('/user/wishlist')
 def show_wishlist():
+    ''' shows events based on a users wishlist'''
+
+    if not g.user:
+        flash('You must be logged in to view this page.', 'danger')
+        return redirect(url_for('homepage'))
+    
     user = User.query.filter_by(username=g.user.username).first()
     wishlist_events = [event for event in user.wishlist]
 
     wishlist_ids = [event.event_id for event in user.wishlist]
+
+    spot_login = True if session['spotify_token'] else False
     
-    return render_template('user-wishlist.html', wishlist=wishlist_events, wishlist_ids=wishlist_ids)
+    
+    return render_template('user-wishlist.html', wishlist=wishlist_events, wishlist_ids=wishlist_ids, user=user, spot_login=spot_login)
 
 
 # --------------- SPOTIFY FLASK ROUTES ---------------
@@ -276,14 +355,21 @@ def show_wishlist():
 
 @app.route('/logout')
 def logout():
-    session['spotify_token'] = None
-    session['top_tracks'] = None
+    ''' logs out user by removing spotify token, current user and top tracks from session'''
+
     do_logout()
     return redirect(url_for('homepage'))
 
 
+
 @app.route('/connect-spotify')
 def login_with_spotify():
+    ''' connects spotify account to currently logged in user. '''
+
+    if not g.user:
+        flash('You must be logged in to view this page.', 'danger')
+        return redirect(url_for('homepage'))
+    
     auth_url = spotify.login_with_spotify()
     if auth_url:
         return redirect(auth_url)
@@ -292,6 +378,12 @@ def login_with_spotify():
 
 @app.route('/switch-accounts')
 def switch_account():
+    ''' switches currently logged in user from currently linked spotify account to another spotify account'''
+
+    if not g.user:
+        flash('You must be logged in to view this page.', 'danger')
+        return redirect(url_for('homepage'))
+    
     auth_url = spotify.swtich_account()
     if auth_url:
         return redirect(auth_url)
@@ -300,6 +392,8 @@ def switch_account():
 
 @app.route('/callback')
 def callback():
+    ''' call back for the spotify API to redirect to after authentication'''
+
     code = request.args.get('code')
     if code:
         info = spotify.callback(code)
@@ -308,6 +402,7 @@ def callback():
             session['spotify_token'] = info
             access_token = spotify.check_refesh_get_token(token_info=info)
             headers = {'Authorization': f'Bearer {access_token}'}
+
             top_artists = spotify.get_cur_u_top_artists(headers)
             session['top_tracks'] = spotify.get_cur_u_top_tracks(headers)
 
@@ -324,24 +419,33 @@ def callback():
 
 @app.route('/top-artists-events')
 def get_top_artists():
+    ''' gets current users top artists events. used to call with front end javascript. groups the events into 2 lists so no artist is on the same list twice'''
 
-    limit = request.args.get('limit', 16)
     if not g.user:
-        return None
+        flash('You must be logged in to view this page.', 'danger')
+        return redirect(url_for('homepage'))
     
     if not session['spotify_token']:
-        return None
+        flash('You must connect Spotify to view this page.', 'danger')
+        return redirect(url_for('homepage'))
+    
+        # gets limit from params
+    limit = request.args.get('limit', 16)
 
     user = User.query.filter_by(username=g.user.username).first()    
 
+        # creates list of users top artists
     artists = user.artists
     if not artists:
         return None
-    
+
+        # gets all events ordered by date connected to a user
     user_events = user.events.order_by(Event.date.asc()).limit(limit).all()
+
     ordered_events = []
     top_events = []
 
+        # if no events already in data base, adds them here 
     if not user_events:      
         zipcode = user.zipcode
         coords = get_lat_long(zipcode)
@@ -363,15 +467,20 @@ def get_top_artists():
         for event_group in all_events:
             if itteration == len(all_events):
                 looped = True
+
+                # if looped once, went through both groups
             elif itteration == len(all_events) * 2:
                 break
 
+                # if not looped yet, appends second event from each group
             if looped:
                 try:
                     ordered_events.append(event_group[1])
                     itteration += 1
                 except IndexError as e:
                     print(f'No index of 2... skipping')
+
+                # if not looped yet, appends first event from each group
             elif not looped:
                 ordered_events.append(event_group[0])
                 itteration += 1
@@ -392,6 +501,7 @@ def get_top_artists():
     x = 0
     i = 2
 
+        # groups events into dous
     for _ in range(round(len(top_events)/2)):
         section = top_events[x:i]
         all_top_events.append(section)
@@ -401,46 +511,38 @@ def get_top_artists():
     return all_top_events if all_top_events else None
 
 
-@app.route('/get-authorization')
-def get_authorization():
-    if g.user:
-        login = True
-    else:
-        login = False
-    if session.get('spotify_token', False):
-        spotify = True
-    else: 
-        spotify = False
-    return {'login': login, 'spotify': spotify}
-    
-
-@app.route('/testing')
-def testing():
-    return render_template('test.html')
-
-
 # ==============================================================
         # PYTHON FUNCTIONS
 # ==============================================================
 
 
 def do_login(user):
+    ''' adds user to current user in session'''
+
     session[CUR_U_ID] = user.id
 
 
 def do_logout():
+    ''' logs out user by deleting all session items'''
+
     if CUR_U_ID in session:
         del session[CUR_U_ID]
 
+    if 'spotify_token' in session:
+        del session['spotify_token']
+
+    if 'top_tracks' in session:
+        del session['top_tracks']
+
 
 def get_lat_long(zipcode):
-    user = {
-        'country_code': 'US',
-        'postal_code': zipcode
-    }
+    ''' gets latitude and longitute based on zipcode '''
 
-    nomi = pgeocode.Nominatim(user.get('country_code', 'US'))
-    data = nomi.query_postal_code(user.get('postal_code', '90001'))
+    user = User.query.filter_by(username=g.user.username).first()
+    country_code = user.country_code
+
+    nomi = pgeocode.Nominatim(country_code)
+    data = nomi.query_postal_code(zipcode)
 
     lat = data.get('latitude', None)
     long = data.get('longitude', None)
@@ -449,6 +551,8 @@ def get_lat_long(zipcode):
 
 
 def get_geohash(coords):
+    ''' gets geohash based on lat and long '''
+
     lat, long = coords
     geohash = pgh.encode(latitude=lat, longitude=long, precision=9)
 
@@ -456,6 +560,8 @@ def get_geohash(coords):
 
 
 def add_artist_to_db(top_artists):
+    ''' adds artist to database if artist not already there'''
+
     u = User.query.filter_by(username=g.user.username).first()
 
     if u.artists:
